requirements.txt
aiofiles==23.2.1
aiogram==3.7.0
aiogram-dialog==2.1.0
aiohttp==3.9.5
aioredis==2.0.1
aioschedule==0.5.2
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
attrs==23.2.0
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
magic-filter==1.0.12
MarkupSafe==2.1.5
multidict==6.0.5
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pydantic==2.7.4
pydantic_core==2.18.4
python-dotenv==1.0.1
redis==5.0.8
requests==2.32.3
SQLAlchemy==2.0.31
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.4
